Extract hightlight transform functionality from highlight pipe
