+++ ng2-components/ng2-alfresco-core/src/pipes/text-highlight.pipe.ts
@@ -15,26 +15,19 @@
  * limitations under the License.
  */
 
-import { Pipe, PipeTransform } from '@angular/core';
+import { Inject, Pipe, PipeTransform } from '@angular/core';
+import { HightlightTransformResult, HighlightTransformService } from '../services/highlight-transform.service';
 
 @Pipe({
     name: 'highlight'
 })
 export class HighlightPipe implements PipeTransform {
 
-    constructor() { }
+    constructor(private highlightTransformService: HighlightTransformService) { }
 
-    transform(text: string, search: string): any {
-        if (search && text) {
-            let pattern = search.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, '\\$&');
-            pattern = pattern.split(' ').filter((t) => {
-                return t.length > 0;
-            }).join('|');
-            const regex = new RegExp(pattern, 'gi');
-            let result = text.replace(regex, (match) => `<span class="highlight">${match}</span>`);
-            return result;
-        } else {
-            return text;
-        }
+    transform(text: string, search: string): string {
+        const HighlightTransform = this.highlightTransformService;
+        const result: HightlightTransformResult = HighlightTransformService.highlight(text, search);
+        return result.text;
     }
 }
